+++ src/lib/recursion_level.rs
@@ -9,20 +9,33 @@ mod recursion_level_test;
 
 use envmnt;
 
-static RECURSION_ENV_VAR_NAME: &str = "CARGO_MAKE_INTERNAL_RECURSION";
+static RECURSION_ENV_VAR_NAME: &str = "CARGO_MAKE_INTERNAL_RECURSION_LEVEL";
 
-pub(crate) fn recursion_level() -> u32 {
-    envmnt::get_or(RECURSION_ENV_VAR_NAME, "0").parse().expect("failed to retrieve env var")
+pub(crate) fn get() -> u32 {
+    let level = envmnt::get_or(RECURSION_ENV_VAR_NAME, "0");
+
+    match level.parse() {
+        Ok(value) => value,
+        Err(error) => {
+            debug!(
+                "Failed to retrieve the recursion level environment variable, error: {}",
+                error
+            );
+            0
+        }
+    }
 }
 
-pub(crate) fn is_first_level() -> bool {
-    recursion_level() == 0
+pub(crate) fn is_top() -> bool {
+    get() == 0
 }
 
-pub(crate) fn increase_level() {
-    envmnt::set(
-        RECURSION_ENV_VAR_NAME,
-        if envmnt::exists(RECURSION_ENV_VAR_NAME) { (recursion_level() + 1).to_string() }
-        else { "0".to_string() }
-    )
+pub(crate) fn increment() {
+    let level = if envmnt::exists(RECURSION_ENV_VAR_NAME) {
+        (get() + 1).to_string()
+    } else {
+        "0".to_string()
+    };
+
+    envmnt::set(RECURSION_ENV_VAR_NAME, level)
 }